chore: StakedTokenBPT size reductions

+++ contracts/z_mocks/governance/MockStakedTokenWithPrice.sol
@@ -11,7 +11,7 @@ import { StakedToken } from "../../governance/staking/StakedToken.sol";
  **/
 contract MockStakedTokenWithPrice is StakedToken {
     /// @notice Most recent PriceCoefficient
-    uint16 public priceCoefficient;
+    uint256 public priceCoefficient;
 
     event PriceCoefficientUpdated(uint16 newPriceCoeff);
 
@@ -55,7 +55,7 @@ contract MockStakedTokenWithPrice is StakedToken {
     /**
      * @dev Get the current priceCoeff
      */
-    function _getPriceCoeff() internal view override returns (uint16) {
+    function _getPriceCoeff() internal view override returns (uint256) {
         return priceCoefficient;
     }
 }
